feat(FN-1861): use common library utilisation report status in all services
