+++ dtfs-central-api/api-tests/mocks/utilisation-reports/utilisation-report-reconciliation-summary.ts
@@ -1,7 +1,7 @@
 import { ObjectId } from 'mongodb';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '@ukef/dtfs2-common';
 import { UtilisationReportReconciliationSummary, UtilisationReportReconciliationSummaryItem } from '../../../src/types/utilisation-reports';
 import { MOCK_BANKS } from '../banks';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../../src/constants/utilisation-report-reconciliation-status';
 
 const MOCK_UTILISATION_REPORT_RECONCILIATION_SUMMARY_ITEM: UtilisationReportReconciliationSummaryItem = {
   reportId: new ObjectId('6598162ce10f06e419170321'),
