+++ dtfs-central-api/src/constants/index.js
@@ -4,7 +4,6 @@ const FACILITIES = require('./facilities');
 const ROUTES = require('./routes');
 const AMENDMENT = require('./amendments');
 const PAYLOAD = require('./payloads');
-const { UTILISATION_REPORT_RECONCILIATION_STATUS } = require('./utilisation-report-reconciliation-status');
 const { UTILISATION_REPORT_HEADERS } = require('./utilisation-report-headers');
 const { CURRENCIES } = require('./currencies');
 
@@ -15,7 +14,6 @@ module.exports = {
   ROUTES,
   AMENDMENT,
   PAYLOAD,
-  UTILISATION_REPORT_RECONCILIATION_STATUS,
   UTILISATION_REPORT_HEADERS,
   CURRENCIES,
 };
