+++ dtfs-central-api/src/repositories/utilisation-reports-repo/update-report-status.test.ts
@@ -1,11 +1,11 @@
 import { ObjectId, WithoutId } from 'mongodb';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS, UtilisationReportReconciliationStatus } from '@ukef/dtfs2-common';
 import { updateManyUtilisationReportStatuses } from './update-report-status';
-import { ReportWithStatus, UtilisationReportReconciliationStatus } from '../../types/utilisation-reports';
+import { ReportWithStatus } from '../../types/utilisation-reports';
 import { UploadedByUserDetails, UtilisationReport } from '../../types/db-models/utilisation-reports';
 import db from '../../drivers/db-client';
 import { MOCK_AZURE_FILE_INFO } from '../../../api-tests/mocks/azure-file-info';
 import * as banksRepo from '../banks-repo';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../constants';
 
 console.error = jest.fn();
