+++ dtfs-central-api/src/types/utilisation-reports.ts
@@ -1,10 +1,8 @@
 import { ObjectId } from 'mongodb';
-import { ValuesOf } from '@ukef/dtfs2-common';
-import { IsoMonthStamp, } from './date';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS, UTILISATION_REPORT_HEADERS } from '../constants';
+import { UtilisationReportReconciliationStatus, ValuesOf } from '@ukef/dtfs2-common';
+import { IsoMonthStamp } from './date';
+import { UTILISATION_REPORT_HEADERS } from '../constants';
 import { Currency } from './currency';
-
-export type UtilisationReportReconciliationStatus = ValuesOf<typeof UTILISATION_REPORT_RECONCILIATION_STATUS>;
 
 export type UtilisationReportReconciliationSummaryItem = {
   reportId: ObjectId;
