+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-report.controller.ts
@@ -12,7 +12,7 @@ export const getUtilisationReportById = async (req: GetUtilisationReportByIdRequ
   const { id } = req.params;
 
   try {
-    const utilisationReport = await UtilisationReportRepo.findOneBy({id: Number(id)});
+    const utilisationReport = await UtilisationReportRepo.findOneBy({ id: Number(id) });
 
     if (!utilisationReport) {
       return res.status(404).send();
