+++ libs/common/src/types/azure-file-info.ts
@@ -1,0 +1,7 @@
+export type AzureFileInfo = {
+  folder: string;
+  filename: string;
+  fullPath: string;
+  url: string;
+  mimetype: string;
+};
