+++ libs/common/src/types/index.ts
@@ -2,3 +2,4 @@ export * from './currency';
 export * from './types-helper';
 export * from './utilisation-reports';
 export * from './date';
+export * from './azure-file-info';
