+++ portal-api/src/constants/index.js
@@ -10,7 +10,6 @@ const { FILE_UPLOAD, FILESHARES } = require('./file-upload');
 const SIGN_IN_LINK = require('./sign-in-link');
 const HTTP_ERROR_CAUSES = require('./http-error-causes');
 const PASSPORT_VALIDATION_RESULTS = require('./passport-validation-results');
-const { UTILISATION_REPORT_RECONCILIATION_STATUS } = require('./utilisation-report-reconciliation-status');
 
 module.exports = {
   FACILITIES,
@@ -26,5 +25,4 @@ module.exports = {
   SIGN_IN_LINK,
   HTTP_ERROR_CAUSES,
   PASSPORT_VALIDATION_RESULTS,
-  UTILISATION_REPORT_RECONCILIATION_STATUS,
 };
