+++ portal-api/src/v1/api.js
@@ -318,7 +318,7 @@ const getUtilisationReports = async (bankId, options) => {
 
 /**
  * @param {number} id
- * @returns {Promise<import('@ukef/dtfs2-common/').UtilisationReportResponseBody>}
+ * @returns {Promise<import('@ukef/dtfs2-common').UtilisationReportResponseBody>}
  */
 const getUtilisationReportById = async (id) => {
   try {
