+++ trade-finance-manager-api/api-tests/v1/utilisation-reports/update-utilisation-report-service.api-test.ts
@@ -1,8 +1,8 @@
 import { ObjectId } from 'bson';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '@ukef/dtfs2-common';
 import app from '../../../src/createApp';
 import createApi from '../../api';
 import testUserCache from '../../api-test-users';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../../src/constants';
 
 const { as } = createApi(app);
