+++ trade-finance-manager-api/src/constants/index.js
@@ -13,7 +13,6 @@ const PAYLOAD = require('./payloads');
 const USER = require('./user');
 const REGEX = require('./regex');
 const { FILESHARES } = require('./fileshares');
-const { UTILISATION_REPORT_RECONCILIATION_STATUS } = require('./utilisation-report-reconciliation-status');
 
 module.exports = {
   FACILITIES,
@@ -32,5 +31,4 @@ module.exports = {
   USER,
   REGEX,
   FILESHARES,
-  UTILISATION_REPORT_RECONCILIATION_STATUS,
 };
