--- trade-finance-manager-api/src/constants/utilisation-report-reconciliation-status.ts
@@ -1,6 +1,0 @@
-export const UTILISATION_REPORT_RECONCILIATION_STATUS = {
-  REPORT_NOT_RECEIVED: 'REPORT_NOT_RECEIVED',
-  PENDING_RECONCILIATION: 'PENDING_RECONCILIATION',
-  RECONCILIATION_IN_PROGRESS: 'RECONCILIATION_IN_PROGRESS',
-  RECONCILIATION_COMPLETED: 'RECONCILIATION_COMPLETED',
-} as const;
