+++ trade-finance-manager-api/src/v1/validation/route-validators/update-report-status-payload-validation/index.test.ts
@@ -2,9 +2,8 @@ import { Request } from 'express';
 import { validationResult } from 'express-validator';
 import { createRequest } from 'node-mocks-http';
 import { ObjectId } from 'bson';
-import { UtilisationReportReconciliationStatus } from '@ukef/dtfs2-common';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS, UtilisationReportReconciliationStatus } from '@ukef/dtfs2-common';
 import { updateReportStatusPayloadValidation } from '.';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../../../constants';
 import { TfmSessionUser } from '../../../../types/tfm-session-user';
 import { ReportWithStatus } from '../../../../types/utilisation-reports';
 import { UpdateUtilisationReportStatusRequestBody } from '../../../controllers/utilisation-reports/update-utilisation-report-status.controller';
