+++ trade-finance-manager-api/src/v1/validation/route-validators/update-report-status-payload-validation/index.ts
@@ -1,5 +1,5 @@
 import { body, checkSchema } from 'express-validator';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../../../../constants';
+import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '@ukef/dtfs2-common';
 import { isValidMongoId } from '../../validateIds';
 
 const VALID_UPDATE_PAYLOAD_STATUSES = [
