+++ trade-finance-manager-ui/package.json
@@ -35,6 +35,7 @@
   "dependencies": {
     "@babel/polyfill": "^7.12.1",
     "@ministryofjustice/frontend": "1.8.0",
+    "@ukef/dtfs2-common": "^1.0.0",
     "ajv": "^8.12.0",
     "axios": "^1.6.7",
     "chance": "^1.1.11",
