+++ trade-finance-manager-ui/server/constants/index.js
@@ -8,7 +8,6 @@ const AMENDMENTS = require('./amendments');
 const DECISIONS = require('./decisions.constant');
 const PARTY = require('./party');
 const { BANK_HOLIDAY_REGION } = require('./bank-holiday-region');
-const { UTILISATION_REPORT_RECONCILIATION_STATUS } = require('./utilisation-report-reconciliation-status');
 
 module.exports = {
   DEAL,
@@ -22,5 +21,4 @@ module.exports = {
   DECISIONS,
   PARTY,
   BANK_HOLIDAY_REGION,
-  UTILISATION_REPORT_RECONCILIATION_STATUS,
 };
