+++ trade-finance-manager-ui/server/types/utilisation-reports.ts
@@ -1,8 +1,5 @@
+import { UtilisationReportReconciliationStatus } from '@ukef/dtfs2-common';
 import { IsoDateTimeStamp, IsoMonthStamp } from './date';
-import { ValuesOf } from './types-helper';
-import { UTILISATION_REPORT_RECONCILIATION_STATUS } from '../constants';
-
-export type UtilisationReportReconciliationStatus = ValuesOf<typeof UTILISATION_REPORT_RECONCILIATION_STATUS>;
 
 export type UtilisationReportReconciliationSummaryItem = {
   reportId: string;
