+++ trade-finance-manager-ui/tsconfig.json
@@ -14,7 +14,7 @@
       "target": "es6",                                     /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
       "lib": ["dom", "esnext"],                                      /* Specify a set of bundled library declaration files that describe the target runtime environment. */
       // "jsx": "preserve",                                /* Specify what JSX code is generated. */
-      // "experimentalDecorators": true,                   /* Enable experimental support for TC39 stage 2 draft decorators. */
+      "experimentalDecorators": true,                   /* Enable experimental support for TC39 stage 2 draft decorators. */
       // "emitDecoratorMetadata": true,                    /* Emit design-type metadata for decorated declarations in source files. */
       // "jsxFactory": "",                                 /* Specify the JSX factory function used when targeting React JSX emit, e.g. 'React.createElement' or 'h' */
       // "jsxFragmentFactory": "",                         /* Specify the JSX Fragment reference used for fragments when targeting React JSX emit e.g. 'React.Fragment' or 'Fragment'. */
